Break Modal into components
